+++ src/components/Dashboard/DashboardItem/AppItem.js
@@ -4,7 +4,7 @@ import { Launch, Info, Favorite, Share, Apps } from "@material-ui/icons";
 
 import { formatDate } from "@cyverse-de/ui-lib";
 
-import * as constants from "../Dashboard/constants";
+import * as constants from "../constants";
 
 import ItemBase, { ItemAction } from "./ItemBase";
